+++ gazebo/rendering/GpuLaser.cc
@@ -17,6 +17,10 @@
 
 #include <sstream>
 
+#include <ignition/math/Helpers.hh>
+#include <ignition/math/Pose3.hh>
+#include <ignition/math/Vector3.hh>
+
 #ifndef _WIN32
   #include <dirent.h>
 #else
@@ -25,10 +29,6 @@
   #include <Winsock2.h>
   #include "gazebo/common/win_dirent.h"
 #endif
-
-#include <ignition/math/Helpers.hh>
-#include <ignition/math/Pose3.hh>
-#include <ignition/math/Vector3.hh>
 
 #include "gazebo/rendering/ogre_gazebo.h"
 
@@ -116,9 +116,9 @@ void GpuLaser::Load()
 void GpuLaser::Init()
 {
   Camera::Init();
-  this->dataPtr->w2nd = this->GetImageWidth();
-  this->dataPtr->h2nd = this->GetImageHeight();
-  this->dataPtr->visual.reset(new Visual(this->GetName()+"second_pass_canvas",
+  this->dataPtr->w2nd = this->ImageWidth();
+  this->dataPtr->h2nd = this->ImageHeight();
+  this->dataPtr->visual.reset(new Visual(this->Name()+"second_pass_canvas",
      this->GetScene()->GetWorldVisual()));
 }
 
@@ -160,7 +160,7 @@ void GpuLaser::CreateLaserTexture(const std::string &_textureName)
     this->dataPtr->firstPassTextures[i] =
       Ogre::TextureManager::getSingleton().createManual(
       texName.str(), "General", Ogre::TEX_TYPE_2D,
-      this->GetImageWidth(), this->GetImageHeight(), 0,
+      this->ImageWidth(), this->ImageHeight(), 0,
       Ogre::PF_FLOAT32_RGB, Ogre::TU_RENDERTARGET).getPointer();
 
     this->Set1stPassTarget(
@@ -179,7 +179,7 @@ void GpuLaser::CreateLaserTexture(const std::string &_textureName)
       _textureName + "second_pass",
       "General",
       Ogre::TEX_TYPE_2D,
-      this->GetImageWidth(), this->GetImageHeight(), 0,
+      this->ImageWidth(), this->ImageHeight(), 0,
       Ogre::PF_FLOAT32_RGB,
       Ogre::TU_RENDERTARGET).getPointer();
 
@@ -583,8 +583,8 @@ void GpuLaser::Set2ndPassTarget(Ogre::RenderTarget *_target)
         GZ_VISIBILITY_ALL & ~(GZ_VISIBILITY_GUI | GZ_VISIBILITY_SELECTABLE));
   }
   Ogre::Matrix4 p = this->BuildScaledOrthoMatrix(
-      0, static_cast<float>(this->GetImageWidth() / 10.0),
-      0, static_cast<float>(this->GetImageHeight() / 10.0),
+      0, static_cast<float>(this->ImageWidth() / 10.0),
+      0, static_cast<float>(this->ImageHeight() / 10.0),
       0.01, 0.02);
 
   this->dataPtr->orthoCam->setCustomProjectionMatrix(true, p);
@@ -600,7 +600,7 @@ void GpuLaser::SetRangeCount(const unsigned int _w, const unsigned int _h)
 /////////////////////////////////////////////////
 void GpuLaser::CreateMesh()
 {
-  std::string meshName = this->GetName() + "_undistortion_mesh";
+  std::string meshName = this->Name() + "_undistortion_mesh";
 
   common::Mesh *mesh = new common::Mesh();
   mesh->SetName(meshName);
@@ -610,7 +610,7 @@ void GpuLaser::CreateMesh()
   double dx, dy;
   submesh->SetPrimitiveType(common::SubMesh::POINTS);
 
-  double viewHeight = this->GetImageHeight()/10.0;
+  double viewHeight = this->ImageHeight()/10.0;
 
   if (this->dataPtr->h2nd == 1)
     dy = 0;
@@ -626,7 +626,7 @@ void GpuLaser::CreateMesh()
 
   double vAngMin = -phi;
 
-  if (this->GetImageHeight() == 1)
+  if (this->ImageHeight() == 1)
     phi = 0;
 
   unsigned int ptsOnLine = 0;
@@ -654,7 +654,7 @@ void GpuLaser::CreateMesh()
       delta = delta - theta;
 
       startX -= dx;
-      if (ptsOnLine == this->GetImageWidth())
+      if (ptsOnLine == this->ImageWidth())
       {
         ptsOnLine = 0;
         startX = 0;