+++ gazebo/sensors/DepthCameraSensor.cc
@@ -104,8 +104,8 @@ void DepthCameraSensor::Init()
     this->camera->Load(cameraSdf);
 
     // Do some sanity checks
-    if (this->camera->GetImageWidth() == 0 ||
-        this->camera->GetImageHeight() == 0)
+    if (this->camera->ImageWidth() == 0 ||
+        this->camera->ImageHeight() == 0)
     {
       gzthrow("image has zero size");
     }
@@ -132,7 +132,7 @@ void DepthCameraSensor::Init()
 void DepthCameraSensor::Fini()
 {
   Sensor::Fini();
-  this->scene->RemoveCamera(this->camera->GetName());
+  this->scene->RemoveCamera(this->camera->Name());
   this->camera.reset();
   this->scene.reset();
 }