+++ Changelog.md
@@ -3,6 +3,9 @@
 1. Use opaque pointers for the rendering/Heightmap class.
     * [Pull request #2074](https://bitbucket.org/osrf/gazebo/pull-request/2074)
 
+1. Deprecate functions in the rendering/Camera class.
+    * [Pull request #2076](https://bitbucket.org/osrf/gazebo/pull-request/2076)
+
 1. Use opaque pointers in the rendering/GpuLaser class.
     * [Pull request #2070](https://bitbucket.org/osrf/gazebo/pull-request/2070)